+++ menu/src/main/java/com/southwind/repository/MenuRepository.java
@@ -8,7 +8,7 @@ public interface MenuRepository {
     public List<Menu> findAll(int index,int limit);
     public int count();
     public void save(Menu menu);
-    public Menu findById(int id);
+    public Menu findById(long id);
     public void update(Menu menu);
-    public void deleteById(int id);
+    public void deleteById(long id);
 }
